requirements.txt

streamlit
opencv-python-headless
pytesseract
Pillow
numpy
